Improves search and artist profile features
Enhances search functionality by adding category mapping and handling Enter key submission.
Moves landing page components to the homepage and refactors artist profile loading and display,
fetching artist posts via artist ID and improving error handling.
Also, fixes an issue where custom images were not being displayed correctly on artist cards.
