+++ frontend/.gitignore
@@ -3,7 +3,7 @@ logs
 *.log
 npm-debug.log*
 yarn-debug.log*
-yarn-error.log*
+# yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
